Title: 학점 계산기, Time: 161ms, Memory: 12MB, Status: Passed - CodeTree

+++ 240130/학점 계산기/credit-calculator.java
@@ -4,13 +4,13 @@ public class Main {
         // 여기에 코드를 작성해주세요.
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        float[] f = new float[n];
-        float sum = 0;
+        double[] f = new double[n];
+        double sum = 0;
         for (int i = 0; i < f.length; i++) {
-            f[i] = sc.nextFloat();
+            f[i] = sc.nextDouble();
             sum += f[i];
         }
-        float avg = sum / f.length;
+        double avg = sum / f.length;
         System.out.printf("%.1f\n", avg);
         if (avg >= 4.0) {
             System.out.print("Perfect");
